requirements.txt
clickhouse-connect
kafka-python
pandas
python-dotenv
boto3
websocket-client
pytest
fastapi
uvicorn
pyarrow
fastparquet